+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * #289 Fix dropdown-trigger container html tag
 * #291 Add configurable closing options for Modal (thanks @jwahdatehagh)
 * #298 Fix mobile detection datepicker for SSR
+* #302 Fix datepicker calendar display when disabled
 
 ## 0.5.1